refactor: 파일 구조 재배치
