+++ src/main/kotlin/com/example/backend/domain/user/dto/request/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.example.backend.dto
+package com.example.backend.domain.user.dto.request
 
 import jakarta.validation.constraints.Email
 import jakarta.validation.constraints.NotBlank
