+++ src/main/kotlin/com/example/backend/domain/user/dto/request/UpdateProfileRequest.kt
@@ -1,4 +1,4 @@
-package com.example.backend.dto
+package com.example.backend.domain.user.dto.request
 
 import jakarta.validation.constraints.Size
