+++ src/main/kotlin/com/example/backend/domain/user/dto/response/AuthResponse.kt
@@ -1,4 +1,4 @@
-package com.example.backend.dto
+package com.example.backend.domain.user.dto.response
 
 data class AuthResponse(
     val token: String,
