+++ src/main/kotlin/com/example/backend/domain/user/entity/enums/UserStatus.kt
@@ -1,4 +1,4 @@
-package com.example.backend.entity
+package com.example.backend.domain.user.entity.enums
 
 enum class
