+++ src/main/kotlin/com/example/backend/global/config/JwtProperties.kt
@@ -1,4 +1,4 @@
-package com.example.backend.config
+package com.example.backend.global.config
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.stereotype.Component
